Medical_Chatbot_Llama2_Pinecone/requirements.txt
pinecone-client
langchain==0.0.225
ctransformers==0.2.5
sentence-transformers==2.2.2
openai